refactor(optimizer): rename api.model package to api.spec (PR #527 review)
Per @abhisheknath2011 review comment 3262769497:
> "Can we change the client side API to api.spec instead of api.model?
> This also aligns with existing services."

Mechanical package rename. The 10 api wire types move from
services/optimizer/.../api/model/ to services/optimizer/.../api/spec/.
No type or signature changes.
