+++ services/optimizer/src/main/java/com/linkedin/openhouse/optimizer/api/spec/OperationStatusDto.java
@@ -1,4 +1,4 @@
-package com.linkedin.openhouse.optimizer.api.model;
+package com.linkedin.openhouse.optimizer.api.spec;
 
 /** Lifecycle states for a table operation recommendation. */
 public enum OperationStatusDto {
