+++ services/optimizer/src/main/java/com/linkedin/openhouse/optimizer/api/spec/TableStatsDto.java
@@ -1,4 +1,4 @@
-package com.linkedin.openhouse.optimizer.api.model;
+package com.linkedin.openhouse.optimizer.api.spec;
 
 import java.time.Instant;
 import java.util.Collections;
