+++ services/optimizer/src/main/java/com/linkedin/openhouse/optimizer/api/spec/TableStatsPayloadDto.java
@@ -1,4 +1,4 @@
-package com.linkedin.openhouse.optimizer.api.model;
+package com.linkedin.openhouse.optimizer.api.spec;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.AllArgsConstructor;
